[MatchingHelper] Fix presumed matching
package pack{
	class B {
		toC : C
	}

	class A {}

	class C inherits A,B {}
}

On this example the MatchingHelper will visit classes in this way:
visit B
	visit C (by toC)
		visit A (by inherits)
		visit B (by inherits)

When visiting A we try to match with the first corresponding super
type of C. It's work since A match itself.

When visiting B we try to match super types of C. So we try first to
match B with A. But we visited B before. This lead to presume B = A.

Checking at least the names should avoid this problem