[Test] Deactivate check overriding for default value of attributes
Melange create an new Ecore merging the original Ecore and the
Aspects.
So for features introduced by the semantic we have one getter from
the Ecore model and one getter from the Aspect (which contains the
default value).
Because of K3 giving the priority to the Ecore getter, we get a null
value from the model instead of the default value from the Aspect